Remove now useless sdk/modules.go file and previous cosmos cmds
